requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
ipython
pytest
pytest-asyncio
httpx
pytest-cov
pytest-mock


